test: steward pass — prune, consolidate ade-code suite
