+++ apps/ade-cli/src/tuiClient/__tests__/Drawer.test.tsx
@@ -3,8 +3,15 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { render } from "ink-testing-library";
 import type { AgentChatSessionSummary } from "../../../../desktop/src/shared/types/chat";
 import type { LaneSummary } from "../../../../desktop/src/shared/types/lanes";
+import type { ChatTerminalSession } from "../../../../desktop/src/shared/types/sessions";
 import { Drawer } from "../components/Drawer";
-import type { ClosedCliSessionSummary } from "../closedCliSessions";
+import {
+  closedCliRightPaneRow,
+  closedCliSessionStatusKind,
+  deriveClosedCliSessions,
+  deriveOpenDrawerSessions,
+  type ClosedCliSessionSummary,
+} from "../closedCliSessions";
 
 function stripAnsi(text: string): string {
   return text.replace(/\u001b\[[0-?]*[ -/]*[@-~]/g, "");
@@ -34,6 +41,31 @@ function lane(id: string, name: string, branchRef: string, createdAt: string, ah
     icon: null,
     tags: [],
     createdAt,
+    ...overrides,
+  };
+}
+
+function terminal(overrides: Partial<ChatTerminalSession> = {}): ChatTerminalSession {
+  return {
+    terminalId: "terminal-1",
+    ptyId: null,
+    chatSessionId: null,
+    laneId: "lane-1",
+    laneName: "Lane 1",
+    title: "Codex CLI",
+    toolType: "codex",
+    goal: null,
+    status: "completed",
+    runtimeState: "exited",
+    active: false,
+    startedAt: "2026-01-01T00:00:00.000Z",
+    endedAt: "2026-01-01T00:01:00.000Z",
+    exitCode: 0,
+    pid: null,
+    resumeCommand: "codex resume terminal-1",
+    resumeMetadata: { provider: "codex", targetKind: "session", targetId: "terminal-1", launch: {} },
+    lastOutputPreview: null,
+    summary: null,
     ...overrides,
   };
 }
@@ -81,6 +113,43 @@ describe("Drawer diff stats", () => {
     expect(frame).not.toContain("+141 / −112");
     expect(frame).not.toContain("-18");
     expect(frame).not.toContain("5m");
+  });
+});
+
+describe("Drawer closed CLI sessions", () => {
+  it.each([
+    ["failed status", { status: "failed", exitCode: 1, runtimeState: "killed" }, "failed"],
+    ["non-zero exit", { status: "completed", exitCode: 2, runtimeState: "exited" }, "failed"],
+    ["user close", { status: "disposed", exitCode: 130, runtimeState: "killed" }, "idle"],
+    ["clean close", { status: "completed", exitCode: 0, runtimeState: "exited" }, "idle"],
+  ] as const)("classifies %s for closed-session glyphs", (_label, overrides, expected) => {
+    const [session] = deriveClosedCliSessions([terminal({
+      terminalId: `t-${_label.replace(/\s+/g, "-")}`,
+      ...overrides,
+    })]);
+
+    expect(session).toBeTruthy();
+    expect(closedCliSessionStatusKind(session!)).toBe(expected);
+    expect(closedCliRightPaneRow(session!, null)).toContain(expected === "failed" ? "✗" : "○");
+  });
+
+  it("filters closed CLI sessions out of the open drawer list", () => {
+    const [closed] = deriveClosedCliSessions([terminal()]);
+    const open: AgentChatSessionSummary = {
+      sessionId: "chat-1",
+      laneId: "lane-1",
+      provider: "codex",
+      model: "gpt-5.5",
+      status: "idle",
+      startedAt: "2026-01-01T00:00:00.000Z",
+      endedAt: null,
+      lastActivityAt: "2026-01-01T00:01:00.000Z",
+      lastOutputPreview: null,
+      summary: null,
+    };
+
+    expect(closed).toBeTruthy();
+    expect(deriveOpenDrawerSessions([open, closed!], [closed!])).toEqual([open]);
   });
 });
